requirements.txt
streamlit
openai-whisper
transformers
torch
ffmpeg-python
plotly
wordcloud
pandas
scikit-learn
nltk
reportlab
python-docx